requirements.txt
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
aniso8601==9.0.1
async-timeout==4.0.3
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
fire==0.7.0
Flask==3.0.3
Flask-RESTful==0.3.10
fontmeta==1.6.1
fonttools==4.54.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==0.10.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mplfinance==0.12.10b0
mplfonts==0.0.8
msgpack==1.1.0
numpy==2.0.2
orjson==3.10.10
oss2==2.19.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pycryptodome==3.21.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-rapidjson==1.20
pytz==2024.2
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
rqdatac==3.1.0
six==1.16.0
termcolor==2.5.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
zstandard==0.23.0
